+++ spyne/model/binary.py
@@ -23,6 +23,11 @@ import os
 import base64
 import tempfile
 
+from base64 import b64encode
+from base64 import b64decode
+from binascii import hexlify
+from binascii import unhexlify
+
 try:
     from cStringIO import StringIO
 except ImportError: # Python 3
@@ -46,6 +51,41 @@ class ByteArray(SimpleModel):
 
     __type_name__ = 'base64Binary'
     __namespace__ = "http://www.w3.org/2001/XMLSchema"
+
+    class Attributes(SimpleModel.Attributes):
+        encoding = None
+        """The binary encoding to use when the protocol does not enforce an
+        encoding for binary data.
+
+        One of (None, 'base64', 'hex')
+        """
+
+    _encoding_handlers = {
+        None: ''.join,
+        'hex': hexlify,
+        'base64': b64encode,
+    }
+
+    _decoding_handlers = {
+        None: lambda x: [x],
+        'hex': unhexlify,
+        'base64': b64decode,
+    }
+
+    def __new__(cls, **kwargs):
+        if 'encoding' in kwargs:
+            v = kwargs['encoding']
+
+            if v in (None, 'base64'):
+                kwargs['type_name'] = 'base64Binary'
+            elif v == 'hex':
+                kwargs['type_name'] = 'hexBinary'
+            else:
+                raise ValueError("'encoding' must be one of: %r" % \
+                                (tuple(ByteArray._encoding.handlers.values()),))
+
+
+        return SimpleModel.__new__(cls, **kwargs)
 
     @classmethod
     @nillable_iterable