+++ spyne/protocol/xml/model.py
@@ -78,7 +78,7 @@ def TBaseFromElement(callable):
 
         if prot.validator is prot.SOFT_VALIDATION and not (
                                             cls.validate_native(cls, retval)):
-            raise ValidationError(element.text)
+            raise ValidationError(retval)
         return retval
 
     return base_from_element
@@ -379,5 +379,5 @@ def unicode_from_element(prot, cls, element):
 
     if prot.validator is prot.SOFT_VALIDATION and not (
                                         cls.validate_native(cls, retval)):
-        raise ValidationError(element.text)
+        raise ValidationError(retval)
     return retval