+++ spyne/server/wsgi.py
@@ -17,7 +17,12 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
 #
 
-"""A server transport uses http as transport, and wsgi as bridge api."""
+
+"""
+A server that uses http as transport via wsgi. It doesn't contain any server
+logic.
+"""
+
 
 import logging
 logger = logging.getLogger(__name__)
@@ -26,10 +31,6 @@ import cgi
 import threading
 import itertools
 
-from spyne.auxproc import process_contexts
-from spyne.model.binary import File
-from spyne.protocol.http import HttpRpc
-
 try:
     from cgi import parse_qs
 except ImportError: # Python 3
@@ -37,23 +38,21 @@ except ImportError: # Python 3
 
 try:
     from werkzeug.formparser import parse_form_data
-except ImportError:
-    parse_form_data = None
+except ImportError, e:
+    def parse_form_data(*args, **kwargs):
+        raise e
 
+from spyne.application import get_fault_string_from_exception
+from spyne.auxproc import process_contexts
+from spyne.error import RequestTooLongError
+from spyne.model.binary import File
+from spyne.model.fault import Fault
+from spyne.protocol.http import HttpRpc
+from spyne.protocol.http import HttpPattern
+from spyne.server.http import HttpBase
 from spyne.server.http import HttpMethodContext
 from spyne.server.http import HttpTransportContext
-
-from spyne.error import RequestTooLongError
-
-from spyne.protocol.http import HttpPattern
 from spyne.util import reconstruct_url
-from spyne.server.http import HttpBase
-
-try:
-    from spyne.protocol.soap.mime import apply_mtom
-except ImportError, e:
-    def apply_mtom(*args, **kwargs):
-        raise e
 
 from spyne.const.ansi_color import LIGHT_GREEN
 from spyne.const.ansi_color import END_COLOR
@@ -61,6 +60,13 @@ from spyne.const.http import HTTP_200
 from spyne.const.http import HTTP_404
 from spyne.const.http import HTTP_405
 from spyne.const.http import HTTP_500
+
+
+try:
+    from spyne.protocol.soap.mime import apply_mtom
+except ImportError, e:
+    def apply_mtom(*args, **kwargs):
+        raise e
 
 
 def _get_http_headers(req_env):
@@ -139,8 +145,10 @@ class WsgiApplication(HttpBase):
             called both from success and error cases.
     '''
 
-    def __init__(self, app, chunked=True):
-        HttpBase.__init__(self, app, chunked)
+    def __init__(self, app, chunked=True,
+                max_content_length=2 * 1024 * 1024,
+                block_length=8 * 1024):
+        HttpBase.__init__(self, app, chunked, max_content_length, block_length)
 
         self._allowed_http_verbs = app.in_protocol.allowed_http_verbs
         self._verb_handlers = {
@@ -254,6 +262,8 @@ class WsgiApplication(HttpBase):
                                                     str(len(ctx.transport.wsdl))
         start_response(HTTP_200, _gen_http_headers(ctx.transport.resp_headers))
 
+        ctx.close()
+
         return [ctx.transport.wsdl]
 
     def handle_error(self, p_ctx, others, error, start_response):
@@ -277,7 +287,7 @@ class WsgiApplication(HttpBase):
             # Report but ignore any exceptions from auxiliary methods.
             logger.exception(e)
 
-        return p_ctx.out_string
+        return itertools.chain(p_ctx.out_string, self.__finalize(p_ctx))
 
     def handle_rpc(self, req_env, start_response):
         initial_ctx = WsgiMethodContext(self, req_env,
@@ -309,7 +319,14 @@ class WsgiApplication(HttpBase):
         if p_ctx.transport.resp_code is None:
             p_ctx.transport.resp_code = HTTP_200
 
-        self.get_out_string(p_ctx)
+        try:
+            self.get_out_string(p_ctx)
+
+        except Exception, e:
+            logger.exception(e)
+            p_ctx.out_error = Fault('Server', get_fault_string_from_exception(e))
+            return self.handle_error(p_ctx, others, p_ctx.out_error, start_response)
+
 
         if isinstance(self.app.out_protocol, HttpRpc) and \
                                                p_ctx.out_header_doc is not None:
@@ -444,7 +461,6 @@ class WsgiApplication(HttpBase):
                             for pk, pv in p_method_descriptor.in_message.\
                                                              _type_info.items():
                                 if pk in r.rule:
-                                    from spyne.model.primitive import String
                                     from spyne.model.primitive import Unicode
                                     from spyne.model.primitive import Decimal