+++ CHANGELOG.rst
@@ -35,6 +35,14 @@ spyne-2.10.0
  * Added as_time_zone support to DateTime. It calls
    ``.astimezone(as_time_zone).replace(tzinfo=None)`` on native values.
 
+spyne-2.9.4
+-----------
+ * Fix for Python 2.6 quirk where any decimal.Decimal() is always less than
+   any ``float()``
+ * Fix missing '/' in WsgiMounter.
+ * Fix confusion in ``spyne.model.primitive.Decimal``'s parameter order.
+ * Add forgotten ``HttpBase`` parameters to ``WsgiApplication``.
+
 spyne-2.9.3
 -----------
  * Fix WsgiApplication choking on empty string return value.